requirements.txt
# Core
fastapi==0.121.3
uvicorn==0.38.0
pydantic[email]==2.10.1
python-multipart==0.0.20
python-dotenv==1.2.1

# Auth & Security
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.5.0
email-validator==2.3.0

# Database
pymongo==4.15.5
motor==3.7.1

# AI / ASR
faster-whisper==1.2.1
av==17.0.0
soundfile==0.13.1
numpy==2.3.5
torch==2.10.0
google-generativeai==0.8.6

# Cloud & External
cloudinary==1.44.1
fastapi-mail==1.4.1
httpx==0.27.0

# Scheduler
APScheduler==3.11.1
